ServerHost lagrer nå tilstand mellom sessions, pretty guud
Flyttet ClientHandler fra ServerHost til sitt eget dokument, opprettet
PropertiesHandler for å inneholde skrive og lesefunksjonene som henter
servertilstand fra fil.
