+++ src/javaClientCode/GUI.java
@@ -12,15 +12,31 @@ public class GUI extends Application{
 	
 	
 	public void start(Stage stage) throws Exception{
+		Stage stage1= new Stage();
+		chat chat = new chat();
+		
 		GUIController ctrl = new GUIController(stage);
 		stage.setScene(ctrl.getScene(0));
+		stage.setTitle("TECKS");
 		stage.show();
+		chat.start(stage1);
 	}
 	
 	
 	public static void main(String[] args){
 		launch();
 		
+	}
+	
+	
+	public class chat extends Application{
+		public void start(Stage stage) throws Exception{
+			Pane root = new Pane();
+			Scene scene = new Scene(root, 200, 200);
+			stage.setScene(scene);
+			stage.setTitle("Chat");
+			stage.show();
+		}
 	}
